Should have correct auth flow Auth --> SideMenu unAuth --> Login
